Added Python scripts to get probability These scripts get the odds of having lands for a spell with a given deck by a given turn.
combinatorics.py
  Uses dynamic programming to enumerate possible hands for a given number of cards drawn from a deck.
density.py
  Calculates the probability of getting any given hand.
  Also includes helper functions to prune combinations by the constraint of satisfying our mana cost.

Depending on how many turns will be supported, it may make sense to rewrite the algorithms in javascript and distribute the workload to the client.
When time allows, it would be nice to substitute the constraint checking in density.py for a full CSP algorithm to allow multicolored lands.